Port to i386

+++ src/include/shmemutil.hpp
@@ -28,7 +28,8 @@ namespace {
 		}
 
 		operator struct timespec() const {
-			return {static_cast<time_t>(this->last_time_won_sec), std::chrono::duration_cast<std::chrono::nanoseconds>(subsec{this->last_time_won_subsec}).count()};
+			return {static_cast<time_t>(this->last_time_won_sec),
+			        static_cast<int>(std::chrono::duration_cast<std::chrono::nanoseconds>(subsec{this->last_time_won_subsec}).count())};
 		}
 
 		static auto map(const char * self, shm_t *& shm) -> int {
